requirements.txt
fastapi>=0.111.0
uvicorn>=0.25.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.10.0
langchain-core>=0.0.0
langchain-groq>=0.0.0
langgraph>=0.0.0
